modify kappa computation
For very noisy data, H.1 becomes quite noisy itself,
with large impact on the kappa, and therefore a prior
that depends on the noise in the data.

The approximation we use now is
`kappa^2 = A^T . ((1 / (A.x + c)) * A.1 )`
i.e. the "central part" of the Hessian `y/(A.x+c)^2` is replaced.
Moreover, we first filter the initial image with a 3D Gaussian with
the given FWHM.

WARNING: this makes the kappa calculation specific to Poisson obj funcs.

WARNING: compute_kappa_image now takes an acq_model as first argument,
no longer the obj_fun